Crowning removed from ifdef ELBERETH
For whatever reason, crowning is only included in games that also include the special effects of engraving Elbereth. My guess is that this is because the lawful crowning title is "the Hand of Elbereth."

This patch removes that dependency. If ELBERETH is not set, lawful characters will be dubed "the Arm of the Law" instead.

This patch does not include any Pirate-specific code.