Fix game corruption/crash bug in tame_song.
Function failed to propperly account for cases where tamedog
refuses to tame the specified monster (for example, if it is
given a human or demon).

This lead to the function accessing an altering a nonexistant
edog struct, which meant that whatever happend to be in memory
after the target monster was being altered in invalid ways,
leading to game instability and save corruption.